requirements.txt
Django==4.2.1
djangorestframework==3.14.0
django-cors-headers==4.0.0
psycopg2-binary==2.9.6

